# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BTM",
        "description": "API du site BTM",
        "contact": {
            "email": "contact@balancetonmatch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/clubs": {
            "put": {
                "tags": [
                    "club"
                ],
                "summary": "Update an existing club",
                "operationId": "updateClub",
                "requestBody": {
                    "$ref": "http://127.0.0.1:8000/components/requestBodies/Club"
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Club not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "clubstore_auth": [
                            "write:clubs",
                            "read:clubs"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "club"
                ],
                "summary": "Add a new club to the store",
                "operationId": "addClub",
                "parameters": [
                    {
                        "name": "abbreviation",
                        "in": "path",
                        "description": "Abbreviation du club",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "http://127.0.0.1:8000/components/requestBodies/Club"
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "clubstore_auth": [
                            "write:clubs",
                            "read:clubs"
                        ]
                    }
                ]
            }
        },
        "/clubs/{clubId}": {
            "get": {
                "tags": [
                    "club"
                ],
                "summary": "Find club by ID",
                "description": "Returns a single club",
                "operationId": "getClubById",
                "parameters": [
                    {
                        "name": "clubId",
                        "in": "path",
                        "description": "ID of club to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/Club"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/Club"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Club not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "club"
                ],
                "summary": "Updates a club in the store with form data",
                "operationId": "updateClubWithForm",
                "parameters": [
                    {
                        "name": "clubId",
                        "in": "path",
                        "description": "ID of club that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the club",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Updated status of the club",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "clubstore_auth": [
                            "write:clubs",
                            "read:clubs"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "club"
                ],
                "summary": "Deletes a club",
                "operationId": "deleteClub",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "clubId",
                        "in": "path",
                        "description": "Club id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Club not found"
                    }
                },
                "security": [
                    {
                        "clubstore_auth": [
                            "write:clubs",
                            "read:clubs"
                        ]
                    }
                ]
            }
        },
        "/clubs/{clubId}/uploadImage": {
            "post": {
                "tags": [
                    "club"
                ],
                "summary": "uploads an image",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "clubId",
                        "in": "path",
                        "description": "ID of club to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Upload images request body",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "clubstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
...
        "/comment/{commentId}/uploadImage": {
            "post": {
                "tags": [
                    "comment"
                ],
                "summary": "uploads an image",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of comment to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Upload images request body",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "commentstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
...
        "/user/{userId}/uploadImage": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "uploads an image",
                "operationId": "uploadFile",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Upload images request body",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "http://127.0.0.1:8000/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "userstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}